+++ lib/main.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
-import './widgets/user_transaction.dart';
+import './widgets/new_transaction.dart';
+import './widgets/transaction_list.dart';
+import './model/transaction.dart';
 
 void main() {
   runApp(const MyApp());
@@ -21,14 +23,41 @@ class MyApp extends StatelessWidget {
   }
 }
 
-class MyHomePage extends StatelessWidget {
+class MyHomePage extends StatefulWidget {
+  @override
+  State<MyHomePage> createState() => _MyHomePageState();
+}
+
+class _MyHomePageState extends State<MyHomePage> {
   // const MyHomePage({super.key});
-
-  // String titleInput;
-  // String amountInput;
-
   final titleController = TextEditingController();
   final amountController = TextEditingController();
+
+  final List<Transaction> _userTransactions = [
+    Transaction(id: 't1', title: 'ランニングシューズ', amount: 18000, date: DateTime.now()),
+    Transaction(id: 't2', title: '１週間分の野菜', amount: 9800, date: DateTime.now())
+  ];
+
+  void _addNewTransaction(String txTitle, double txAmount) {
+    final newTx = Transaction(id: DateTime.now().toString(), title: txTitle, amount: txAmount, date: DateTime.now());
+
+    setState(() {
+      _userTransactions.add(newTx);
+    });
+  }
+
+  void _startAddNewTransaction(BuildContext ctx) {
+    showModalBottomSheet(context: ctx, builder: (_) {
+      return GestureDetector(
+        onTap: () {
+          
+        },
+        child: NewTransaction(addTransactionHandler: _addNewTransaction),
+        behavior: HitTestBehavior.opaque,
+      );
+    
+    });
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -36,7 +65,7 @@ class MyHomePage extends StatelessWidget {
       appBar: AppBar(
         title: Text('Flutter App'),
         actions: [
-          IconButton(onPressed: null, icon: Icon(Icons.add))
+          IconButton(onPressed: () => _startAddNewTransaction(context), icon: Icon(Icons.add))
         ],
       ),
       body: SingleChildScrollView(
@@ -54,14 +83,14 @@ class MyHomePage extends StatelessWidget {
                 elevation: 5,
               ),
             ),
-            UserTransaction()
+            TransactionList(transactions: _userTransactions),
           ],
         ),
       ),
       floatingActionButtonLocation: FloatingActionButtonLocation.centerFloat,
       floatingActionButton: FloatingActionButton(
         child:Icon(Icons.add),
-        onPressed: null,
+        onPressed: () => _startAddNewTransaction(context),
       ),
     );
   }
