+++ lib/widgets/new_transaction.dart
@@ -1,22 +1,30 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 
-class NewTransaction extends StatelessWidget {
+class NewTransaction extends StatefulWidget {
   // const NewTransaction({super.key});
   final void Function(String, double) addTransactionHandler;
 
-  final titleController = TextEditingController();
-  final amountController = TextEditingController();
+
   NewTransaction({
     required this.addTransactionHandler,
     Key? key,
   }): super(key: key);
 
+  @override
+  State<NewTransaction> createState() => _NewTransactionState();
+}
+
+class _NewTransactionState extends State<NewTransaction> {
+  final titleController = TextEditingController();
+
+  final amountController = TextEditingController();
+
   void submitData() {
     final enteredTitle = titleController.text;
     final enteredAmount = double.parse(amountController.text);
     if(enteredTitle.isEmpty || enteredAmount <= 0) return;
-    addTransactionHandler(titleController.text, double.parse(amountController.text));
+    widget.addTransactionHandler(titleController.text, double.parse(amountController.text));
+    Navigator.of(context).pop();
   }
 
   @override
